requirements.txt
Django==4.2.11
djangorestframework==3.15.2
openpyxl==3.1.5
qrcode==7.4.2
Pillow==10.3.0
